+++ app/src/main/java/com/example/kdiakonidze/beerapeni/models/ShekvetebiGR.java
@@ -11,10 +11,10 @@ public class ShekvetebiGR implements Serializable{
     private String name;
     private ArrayList<Shekvetebi> childs;
     private ArrayList<ShekvetebiSum> grHeadOrderSum = new ArrayList<>();;
-    private int k30w = 0;
-    private int k50w = 0;
-    private int k30 = 0;
-    private int k50 = 0;
+    private float k30w = 0;
+    private float k50w = 0;
+    private float k30 = 0;
+    private float k50 = 0;
 
     public ShekvetebiGR(String name, ArrayList<Shekvetebi> childs) {
         this.name = name;
@@ -37,35 +37,35 @@ public class ShekvetebiGR implements Serializable{
         this.childs = childs;
     }
 
-    public int getK30w() {
+    public float getK30w() {
         return k30w;
     }
 
-    public void setK30w(int k30w) {
+    public void setK30w(float k30w) {
         this.k30w = k30w;
     }
 
-    public int getK50w() {
+    public float getK50w() {
         return k50w;
     }
 
-    public void setK50w(int k50w) {
+    public void setK50w(float k50w) {
         this.k50w = k50w;
     }
 
-    public int getK30() {
+    public float getK30() {
         return k30;
     }
 
-    public void setK30(int k30) {
+    public void setK30(float k30) {
         this.k30 = k30;
     }
 
-    public int getK50() {
+    public float getK50() {
         return k50;
     }
 
-    public void setK50(int k50) {
+    public void setK50(float k50) {
         this.k50 = k50;
     }
